requirements.txt
alembic==1.14.0
autopep8==2.3.1
blinker==1.9.0
click==8.1.8
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
pytesseract==0.3.13
SQLAlchemy==2.0.36
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
